refactor(wallet): split LedgerTable component out of WalletPanel
