+++ frontend/src/components/wallet/WalletPanel.tsx
@@ -1,115 +1,22 @@
 "use client";
 
-import { useState, type ComponentType } from "react";
+import { useState } from "react";
 import { useQueryClient } from "@tanstack/react-query";
 import { Button } from "@/components/ui/Button";
 import { Input } from "@/components/ui/Input";
 import { LoadingSpinner } from "@/components/ui/LoadingSpinner";
 import { Modal } from "@/components/ui/Modal";
-import {
-  AlertTriangle,
-  ArrowDownToLine,
-  ArrowUpFromLine,
-  Clock,
-  Lock,
-  Pencil,
-  Tag,
-  Undo2,
-  Unlock,
-  Wallet,
-} from "@/components/ui/icons";
-import { cn } from "@/lib/cn";
+import { AlertTriangle } from "@/components/ui/icons";
 import {
   withdraw,
   payPenalty,
   acceptTerms,
 } from "@/lib/api/wallet";
 import { useWallet, walletQueryKey } from "@/lib/wallet/use-wallet";
-import { formatRelativeTime } from "@/lib/time/relativeTime";
-import type { LedgerEntry } from "@/types/wallet";
+import { LedgerTable } from "@/components/wallet/LedgerTable";
 
 function formatLindens(amount: number): string {
   return `L$${amount.toLocaleString()}`;
-}
-
-function entryTypeLabel(t: LedgerEntry["entryType"]): string {
-  switch (t) {
-    case "DEPOSIT": return "Deposit";
-    case "WITHDRAW_QUEUED": return "Withdraw queued";
-    case "WITHDRAW_COMPLETED": return "Withdraw completed";
-    case "WITHDRAW_REVERSED": return "Withdraw reversed";
-    case "BID_RESERVED": return "Bid reserved";
-    case "BID_RELEASED": return "Bid released";
-    case "ESCROW_DEBIT": return "Escrow funded";
-    case "ESCROW_REFUND": return "Escrow refund";
-    case "LISTING_FEE_DEBIT": return "Listing fee paid";
-    case "LISTING_FEE_REFUND": return "Listing fee refund";
-    case "PENALTY_DEBIT": return "Penalty paid";
-    case "ADJUSTMENT": return "Adjustment";
-  }
-}
-
-type EntryVisual = {
-  Icon: ComponentType<{ className?: string }>;
-  tone: string;
-};
-
-/**
- * Maps ledger entry types to a lucide icon + a Material-3 colour token
- * pair. Inflows (deposit / refund) use {@code text-success}; outflows that
- * are user-initiated and final use the neutral {@code text-on-surface};
- * "in-progress" or warning-tinted entries (queued withdraw, reserved bid,
- * penalty) use {@code text-warning}.
- */
-function entryVisual(t: LedgerEntry["entryType"]): EntryVisual {
-  switch (t) {
-    case "DEPOSIT":
-      return { Icon: ArrowDownToLine, tone: "text-success" };
-    case "WITHDRAW_QUEUED":
-      return { Icon: Clock, tone: "text-on-surface-variant" };
-    case "WITHDRAW_COMPLETED":
-      return { Icon: ArrowUpFromLine, tone: "text-on-surface" };
-    case "WITHDRAW_REVERSED":
-      return { Icon: Undo2, tone: "text-warning" };
-    case "BID_RESERVED":
-      return { Icon: Lock, tone: "text-warning" };
-    case "BID_RELEASED":
-      return { Icon: Unlock, tone: "text-on-surface-variant" };
-    case "ESCROW_DEBIT":
-      return { Icon: ArrowUpFromLine, tone: "text-on-surface" };
-    case "ESCROW_REFUND":
-      return { Icon: ArrowDownToLine, tone: "text-success" };
-    case "LISTING_FEE_DEBIT":
-      return { Icon: Tag, tone: "text-on-surface" };
-    case "LISTING_FEE_REFUND":
-      return { Icon: ArrowDownToLine, tone: "text-success" };
-    case "PENALTY_DEBIT":
-      return { Icon: AlertTriangle, tone: "text-warning" };
-    case "ADJUSTMENT":
-      return { Icon: Pencil, tone: "text-on-surface-variant" };
-  }
-}
-
-/**
- * Recent-ledger row date label. Within the last 24 hours we render
- * a relative string ({@code "3m ago"}); older entries fall back to a
- * compact absolute date-time. Title attribute always holds the
- * absolute-locale string so a hover surfaces the precise instant.
- */
-function formatLedgerDate(iso: string): string {
-  const d = new Date(iso);
-  if (!Number.isFinite(d.getTime())) return iso;
-  const hoursAgo = (Date.now() - d.getTime()) / (1000 * 60 * 60);
-  if (hoursAgo < 24) {
-    return formatRelativeTime(d);
-  }
-  return d.toLocaleString(undefined, {
-    year: "numeric",
-    month: "short",
-    day: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-  });
 }
 
 function genIdempotencyKey(): string {
@@ -224,58 +131,7 @@ export function WalletPanel() {
         <h3 className="text-lg font-semibold text-on-surface mb-3">
           Recent Activity
         </h3>
-        {wallet.recentLedger.length === 0 ? (
-          <div className="flex flex-col items-center text-center py-8 gap-3">
-            <Wallet className="h-12 w-12 text-on-surface-variant/40" />
-            <h3 className="font-medium text-on-surface">No activity yet</h3>
-            <p className="text-sm text-on-surface-variant max-w-sm">
-              Visit any SLPA Terminal in-world to make your first deposit.
-              Locations: SLPA HQ and partner auction venues.
-            </p>
-          </div>
-        ) : (
-          <ul className="text-sm">
-            {wallet.recentLedger.map((e) => {
-              const { Icon, tone } = entryVisual(e.entryType);
-              return (
-                <li
-                  key={e.id}
-                  className={cn(
-                    "flex justify-between items-start gap-3 px-2 py-2 rounded-md",
-                    "border-b border-outline-variant last:border-b-0",
-                    "hover:bg-surface-container-low",
-                  )}
-                >
-                  <div className="flex items-start min-w-0">
-                    <Icon
-                      className={cn("h-4 w-4 mr-2 mt-0.5 shrink-0", tone)}
-                      aria-hidden="true"
-                    />
-                    <div className="min-w-0">
-                      <div className="font-medium text-on-surface truncate">
-                        {entryTypeLabel(e.entryType)}
-                      </div>
-                      <div
-                        className="text-xs text-on-surface-variant"
-                        title={new Date(e.createdAt).toLocaleString()}
-                      >
-                        {formatLedgerDate(e.createdAt)}
-                      </div>
-                    </div>
-                  </div>
-                  <div className="text-right shrink-0">
-                    <div className={cn("tabular-nums", tone)}>
-                      {formatLindens(e.amount)}
-                    </div>
-                    <div className="text-xs text-on-surface-variant tabular-nums">
-                      Bal {formatLindens(e.balanceAfter)} / Res {formatLindens(e.reservedAfter)}
-                    </div>
-                  </div>
-                </li>
-              );
-            })}
-          </ul>
-        )}
+        <LedgerTable entries={wallet.recentLedger} />
       </div>
 
       <Modal
